feat: refactor authentication components and implement client-side handling for login, logout, and password testing
